Calculates meshlets for all meshes loaded
