+++ LambdaEngine/Include/Resources/Mesh.h
@@ -36,19 +36,38 @@ namespace LambdaEngine
 	{
 		using IndexType = uint32;
 
-		~Mesh()
+		inline Mesh()
 		{
-			SAFEDELETE_ARRAY(pVertexArray);
-			SAFEDELETE_ARRAY(pIndexArray);
-			SAFEDELETE_ARRAY(pMeshletArray);
 		}
 
-		Vertex* pVertexArray	= nullptr;
-		IndexType* pIndexArray	= nullptr;
-		Meshlet* pMeshletArray	= nullptr;
-		uint32 VertexCount		= 0;
-		uint32 IndexCount		= 0;
-		uint32 MeshletCount		= 0;
+		inline ~Mesh()
+		{
+			SAFEDELETE_ARRAY(pVertexArray);
+			VertexCount = 0;
+
+			SAFEDELETE_ARRAY(pIndexArray);
+			IndexCount = 0;
+
+			SAFEDELETE_ARRAY(pMeshletArray);
+			MeshletCount = 0;
+
+			SAFEDELETE_ARRAY(pUniqueIndices);
+			UniqueIndexCount = 0;
+			
+			SAFEDELETE_ARRAY(pPrimitiveIndices);
+			PrimitiveIndexCount = 0;
+		}
+
+		Vertex*		pVertexArray		= nullptr;
+		IndexType*	pIndexArray			= nullptr;
+		IndexType*	pUniqueIndices		= nullptr;
+		IndexType*	pPrimitiveIndices	= nullptr;
+		Meshlet*	pMeshletArray		= nullptr;
+		uint32		VertexCount			= 0;
+		uint32		IndexCount			= 0;
+		uint32		MeshletCount		= 0;
+		uint32		UniqueIndexCount	= 0;
+		uint32		PrimitiveIndexCount	= 0;
 	};
 
 	class MeshFactory
