+++ LambdaEngine/Source/Launch/Main.cpp
@@ -1,8 +1,6 @@
 #include "Engine/EngineLoop.h"
+
 #include "Debug/Profiler.h"
-
-#include "Application/API/Events/EventQueue.h"
-
 
 namespace LambdaEngine
 {
